Moving identical functions to fix.py
